Add classes for OOP
Extract methods from Duke to create other methods. Yet to extract content for Parser class.
